update no-arg function prototypes to (void)
This doesn't change behaviour, but I'm parsing this headerfile with libclang to generate language bindings for another language, and it doesn't like the `()` no-arg prototypes. This change solves the problem.

According to [SO](https://stackoverflow.com/questions/51032/is-there-a-difference-between-foovoid-and-foo-in-c-or-c), `(void)` is the proper way to specify a function taking no arguments.